fix golangcl-lint Error

+++ compatibility/async/go-client/pkg/user.go
@@ -59,7 +59,7 @@ func (g Gender) String() string {
 		return s
 	}
 
-	return strconv.Itoa(int(g))
+	return strconv.Itoa(int(int32(g)))
 }
 
 func (g Gender) EnumValue(s string) hessian.JavaEnum {
